requirements.txt
astroid==2.4.1
autopep8==1.5.2
Django==2.2.7
django-cors-headers==3.2.1
djangorestframework==3.10.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==6.2.1
pycodestyle==2.6.0
pylint==2.5.2
pytz==2019.3
six==1.15.0
sqlparse==0.3.0
toml==0.10.1
wrapt==1.12.1
